RoboSkctchPortraits/requirements.txt
alabaster==0.7.13
altgraph==0.17.3
Babel==2.12.1
certifi==2023.7.22
charset-normalizer==3.2.0
colorama==0.4.6
customtkinter==5.2.0
darkdetect==0.8.0
docutils==0.18.1
idna==3.4
imagesize==1.4.1
importlib-metadata==6.8.0
install==1.3.5
Jinja2==3.1.2
MarkupSafe==2.1.3
numpy==1.25.2
opencv-python==4.8.0.76
packaging==23.1
pefile==2023.2.7
Pillow==10.0.0
Pygments==2.16.1
pyinstaller==5.13.2
pyinstaller-hooks-contrib==2023.8
pyrealsense2==2.54.1.5217
pywin32-ctypes==0.2.2
requests==2.31.0
snowballstemmer==2.2.0
Sphinx==6.2.1
sphinx-rtd-theme==1.2.2
sphinxcontrib-applehelp==1.0.7
sphinxcontrib-devhelp==1.0.5
sphinxcontrib-htmlhelp==2.0.4
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-plantuml==0.25
sphinxcontrib-qthelp==1.0.6
sphinxcontrib-serializinghtml==1.1.8
urllib3==2.0.4
zipp==3.16.2
